requirements.txt
boto3==1.34.34
python-jose==3.3.0
cryptography==42.0.2
dnspython==2.4.2 